test(core): split plugin sourcemap tests
