+++ packages/core/test/core/plugin-sourcemap/helpers.ts
@@ -1,0 +1,19 @@
+export {
+  describe,
+  test,
+  expect,
+  build,
+  watch,
+  mkdtempSync,
+  writeFileSync,
+  readFileSync,
+  rmSync,
+  join,
+  tmpdir,
+  expectPluginDiagnostic,
+  lineOffsetMappings,
+  parseBundleMap,
+  expectMarkerMappedToSourceLine,
+} from '../helpers';
+
+export type { ZntcPlugin } from '../helpers';
